% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandungjuara-package.R
\docType{package}
\name{bandungjuara}
\alias{bandungjuara}
\alias{bandungjuara-package}
\title{paket \code{bandungjuara}}
\description{
Antarmuka untuk Open Data Bandung
}
\details{
See the README on
\href{https://github.com/aswansyahputra/bandungjuara#readme}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aswansyahputra/bandungjuara}
  \item Report bugs at \url{https://github.com/aswansyahputra/bandungjuara/issues}
}

}
\author{
\strong{Maintainer}: Muhammad Aswan Syahputra \email{aswansyahputra@sensolution.id}

}
