% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cari.R
\name{cari}
\alias{cari}
\title{Cari dataset}
\source{
Dataset bersumber dari Open Data Kota Bandung \url{http://data.bandung.go.id}.
}
\usage{
cari(kata_kunci = "bandung")
}
\arguments{
\item{kata_kunci}{kata kunci dataset yang ingin dicari, case insensitive}
}
\value{
dataframe dengan 5 kolom, yaitu:
\describe{
  \item{nama}{Judul dataset}
  \item{deskripsi}{Penjelasan mengenai dataset}
  \item{url}{url lokasi dataset di peladen}
  \item{dibuat}{Tanggal dataset dibuat atau diunggah, dalam format YYYY-MM-DD}
  \item{diperbaharui}{Tanggal dataset terakhir diperbaharui, dalam format YYYY-MM-DD}
}
}
\description{
Fungsi untuk mencari dataset yang tersedia di Open Data Bandung. Saat ini URL dataset yang ditampilkan hanya yang berasal dari peladen data.bandung.go.id.
}
\examples{

library(bandungjuara)

kebakaran <- cari(kata_kunci = "kebakaran")
kebakaran

}
